// whatsapp-clone-ui/src/app/services/models/conversation-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ConversationResponse {
  id?: string;
  lastMessage?: string;
  lastMessageId?: number;
  lastMessageStatus?: 'PENDING' | 'SENT' | 'DELIVERED' | 'READ' | 'FAILED';
  lastMessageTime?: string;
  lastMessageType?: 'TEXT' | 'IMAGE' | 'VIDEO' | 'AUDIO' | 'FILE';
  name?: string;
  otherUserId?: string;
  otherUserLastSeen?: string;
  otherUserOnline?: boolean;
  type?: 'PRIVATE' | 'GROUP';
  unreadCount?: number;
}