// whatsapp-clone-ui/src/app/services/models/message-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AttachmentResponse } from '../models/attachment-response';
export interface MessageResponse {
  attachments?: Array<AttachmentResponse>;
  content?: string;
  createdAt?: string;
  deleted?: boolean;
  edited?: boolean;
  editedAt?: string;
  id?: number;
  replyToContent?: string;
  replyToMessageId?: number;
  replyToType?: 'TEXT' | 'IMAGE' | 'VIDEO' | 'AUDIO' | 'FILE';
  senderId?: string;
  status?: 'PENDING' | 'SENT' | 'DELIVERED' | 'READ' | 'FAILED';
  type?: 'TEXT' | 'IMAGE' | 'VIDEO' | 'AUDIO' | 'FILE';
}